move readenv to utils
